flask-postgres-app/backend/requirements.txt
Flask==2.2.2
psycopg2-binary==2.9.5
Flask-Cors==4.0.0
Werkzeug==2.2.2